# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"project","description":"project","contact":{"name":"Federico","email":"federico.gonzalez.it@gmail.com"},"version":"1"},"servers":[{"url":"http://localhost:8081","description":"project"}],"tags":[{"name":"Plant","description":"Endpoints for plant management"}],"paths":{"/plant/updatePlant":{"put":{"tags":["Plant"],"operationId":"updatePlant","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetPlantDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlantDto"}}}}}}},"/plant/newPlant":{"post":{"tags":["Plant"],"operationId":"createPlant","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewPlantDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlantDto"}}}}}}},"/auth/register":{"post":{"tags":["auth-controller"],"summary":"Register new user","description":"Creates a new user account","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserDto"}}},"required":true},"responses":{"200":{"description":"User registered successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}},"400":{"description":"Invalid request data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}}}}},"/auth/refresh":{"post":{"tags":["auth-controller"],"summary":"Refresh token","description":"Generates a new access token using a refresh token. Add 'Bearer ' prefix before the token.","operationId":"refreshToken","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Missing or malformed Authorization header","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}},"200":{"description":"Token refreshed successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}},"401":{"description":"Invalid or expired token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"summary":"Login user","description":"Authenticates a user and returns a JWT token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}},"required":true},"responses":{"200":{"description":"User logged in successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}},"400":{"description":"Invalid request data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}},"404":{"description":"User not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}}}}},"/plant/logicDown/{plantId}":{"patch":{"tags":["Plant"],"operationId":"logicDownPlant","parameters":[{"name":"plantId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlantDto"}}}}}}},"/plant/getPlants":{"get":{"tags":["Plant"],"operationId":"getPlants","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetPlantFlagDto"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"GetPlantDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"readings":{"type":"integer","format":"int32"},"medAlerts":{"type":"integer","format":"int32"},"redAlerts":{"type":"integer","format":"int32"},"sensorsDisabled":{"type":"integer","format":"int32"},"country":{"type":"string"},"status":{"type":"boolean"}}},"PlantDto":{"type":"object","properties":{"name":{"type":"string"},"status":{"type":"boolean"}}},"NewPlantDto":{"type":"object","properties":{"name":{"type":"string"},"readings":{"type":"integer","format":"int32"},"medAlerts":{"type":"integer","format":"int32"},"redAlerts":{"type":"integer","format":"int32"},"sensorsDisabled":{"type":"integer","format":"int32"},"country":{"type":"string"}}},"RegisterUserDto":{"required":["email","name","password"],"type":"object","properties":{"name":{"maxLength":50,"minLength":3,"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"},"email":{"type":"string"}}},"TokenResponseDto":{"type":"object","properties":{"access_token":{"type":"string"},"refresh_token":{"type":"string"}}},"LoginUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"GetPlantFlagDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"readings":{"type":"integer","format":"int32"},"medAlerts":{"type":"integer","format":"int32"},"redAlerts":{"type":"integer","format":"int32"},"sensorsDisabled":{"type":"integer","format":"int32"},"country":{"type":"string"},"flag":{"type":"string"},"status":{"type":"boolean"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}