# src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/auth/LoginUserDto.java
package ar.edu.utn.frc.tup.lc.iv.dtos.common.auth;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class LoginUserDto {
    private String email;
    private String password;
}