# src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/auth/TokenResponseDto.java
package ar.edu.utn.frc.tup.lc.iv.dtos.common.auth;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class TokenResponseDto {
    @JsonProperty("access_token")
    private String accessToken;
    @JsonProperty("refresh_token")
    private String refreshToken;
}